requirements.txt
pyyaml==6.0.1
Flask==3.0.2
mysql-connector-python
requests==2.31.0